% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdif_data.R
\docType{data}
\name{cdif_snp_allele_mat}
\alias{cdif_snp_allele_mat}
\title{SNP Allele matrix for the Cdif trehalose/severity data set}
\format{
A matrix with 10,000 rows and 25 columns. Each row name corresponds to a
  variant. Each column corresponds to a sample.
\describe{
  \item{snp_allele_mat}{The letter in the allele matrix indicates the
  variant in that sample. This particular data matrix is a subset of a matrix
  from: /nfs/turbo/umms-esnitkin/Project_Cdiff/Sequence_data/Project_propensity_score_match/consensus/2020_12_07_16_25_20_core_results/data_matrix/matrices}
 }
}
\usage{
cdif_snp_allele_mat
}
\description{
A subset SNP allele matrix to use as an example
}
\keyword{datasets}
