% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdif_data.R
\docType{data}
\name{cdif_tree}
\alias{cdif_tree}
\title{Tree for the Cdif trehalose/severity data set}
\format{
Ape tree format
\describe{
  \item{cdif_tree}{This particular tree  is a subset of a tree from:
  /nfs/turbo/umms-esnitkin/Project_Cdiff/Analysis/Propensity_paper/clinical_cdifficile_trehalose_variants/data/inputs/input_tree.tree }
 }
}
\usage{
cdif_tree
}
\description{
A subset tree to use as an example
}
\keyword{datasets}
