% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{check_ref_choice}
\alias{check_ref_choice}
\title{Check that reference choices makes sense
Issue warning to user if they are not compatible}
\usage{
check_ref_choice(ref_to_anc, ref_to_maj, tree)
}
\arguments{
\item{ref_to_anc}{logical}

\item{ref_to_maj}{logical}

\item{tree}{phylogenetic tree}
}
\description{
Check that reference choices makes sense
Issue warning to user if they are not compatible
}
