% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_clustering_on_tree.R
\name{check_tree_clustering}
\alias{check_tree_clustering}
\title{Check for clustering on tree}
\usage{
check_tree_clustering(
  tree,
  isolate_labels,
  nperm = 1000,
  control_labels = NULL,
  plot_path = "cluster.pdf"
)
}
\arguments{
\item{tree}{Tree to look for clustering on.}

\item{isolate_labels}{Named vector of labels by which pure clusters are defined. Names must be equivalent to tree tip label names.}

\item{nperm}{Number of permutations to perform.}

\item{control_labels}{Named vector of labels known to cluster. Names must be equivalent to tree tip label names. This controls for clustering by requiring that the pure clusters must contain multiple of the control labels.}

\item{plot_path}{Path to output histogram. No plot created if NA.}
}
\value{
Vector of real (1st element) and permuted (all other elements) pure cluster sizes.
}
\description{
This function is used to test for clustering of a certain epi factor on a phylogenetic tree.
}
\examples{
tree = ape::rtree(50)
labs = sample(c(0,1),50,replace=TRUE)
names(labs) = tree$tip.label
check_tree_clustering(tree,labs,nperm=10)
}
