% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{convert_code_to_binary}
\alias{convert_code_to_binary}
\title{Convert the code in the code matrix to either 1s or 0s}
\usage{
convert_code_to_binary(bin_mat)
}
\arguments{
\item{bin_mat}{Matrix}
}
\value{
Matrix
}
\description{
Convert the code in the code matrix to either 1s or 0s
}
