% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greatlakes_resource_usage.R
\name{ConvertTimeToHours}
\alias{ConvertTimeToHours}
\title{Convert sbatch job run time into hours}
\usage{
ConvertTimeToHours(timeStr)
}
\arguments{
\item{timeStr}{String of the form: 1-10:30:00 or 10:30:00. The number before
the dash indicates days.}
}
\value{
Number of hours rounded to 3 digits. Numeric.
}
\description{
the `$ seff <jobid>` command on Great Lakes returns a string
  that includes the run time of an sbatch job. These strings look like:
  * Job Wall-clock time: 10:30:00
  or
  * Job Wall-clock time: 1-10:30:00
  and indicate 10.5 hours or 34.5 hours, respectively. Given such a string
  this function converts the number portion into just hours. The time is
  recorded as hours:minutes:seconds.
}
\examples{
ConvertTimeToHours("1-10:30:00")
ConvertTimeToHours("10:30:00")
}
