% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{define_reference_alleles}
\alias{define_reference_alleles}
\title{Define reference alleles
If you're planning to return a binary matrix you need to pick a reference
allele for each variant. You can choose the ancestral allele, major allele,
or reference allele.}
\usage{
define_reference_alleles(
  return_binary_matrix,
  ref_to_anc,
  ref_to_maj,
  tree,
  varmat_allele,
  major_alleles,
  parallelization
)
}
\arguments{
\item{return_binary_matrix}{logical}

\item{ref_to_anc}{logical}

\item{ref_to_maj}{logical}

\item{tree}{Phylogenetic tree}

\item{varmat_allele}{Matrix}

\item{major_alleles}{Vector}
}
\value{
alleles = character vector, length = nrow(varmat_allele) if
        ref_to_maj==TRUE or ref_to_maj==FALSE & ref_to_anc==FALSE.
        alleles = matrix. Dim = nrow(varmat_allele) x 2 if ref_to_anc==TRUE.
        alleles = NULL if return_binary_matrix==FALSE.
}
\description{
Define reference alleles
If you're planning to return a binary matrix you need to pick a reference
allele for each variant. You can choose the ancestral allele, major allele,
or reference allele.
}
