% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_annotation_info}
\alias{get_annotation_info}
\title{Get SNP or INDEL variant information from matrix row annotations}
\usage{
get_annotation_info(mat_type, mat)
}
\arguments{
\item{mat_type}{Character. Either "SNP" or "INDEL"}

\item{mat}{Varmat code.}
}
\value{
annot_info (list?)
}
\description{
Get SNP or INDEL variant information from matrix row annotations
}
