% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{keep_sites_based_on_conf_logical}
\alias{keep_sites_based_on_conf_logical}
\title{Determine which variants to keep based on user input confidence logical}
\usage{
keep_sites_based_on_conf_logical(bin_mat, logical)
}
\arguments{
\item{bin_mat}{Variant binary matrix.}

\item{logical}{True or false}
}
\value{
logical vector
}
\description{
Determine which variants to keep based on user input confidence logical
}
