% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{load_if_path}
\alias{load_if_path}
\title{Load matrix from path if needed}
\usage{
load_if_path(mat)
}
\arguments{
\item{mat}{- loaded in data.frame of varmat or character string of a path to a varmat}
}
\value{
variant matrix
}
\description{
Loads variant matrix from path if not already loaded
}
