% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_snps.R
\name{parse_snps}
\alias{parse_snps}
\title{Parse SNP matrix from Ali's pipeline}
\usage{
parse_snps(
  varmat_code,
  varmat_allele,
  tree = NULL,
  og = NULL,
  remove_multi_annots = FALSE,
  return_binary_matrix = TRUE,
  ref_to_anc = TRUE,
  keep_conf_only = TRUE,
  mat_suffix = "R1_001.fastq.gz|_R1.fastq.gz"
)
}
\arguments{
\item{varmat_code}{Loaded data.frame or path to the varmat_code file
generated from internal variant calling pipeline}

\item{varmat_allele}{Loaded data.frame or path to the varmat_allele file
generated from internal variant calling pipeline}

\item{tree}{Optional: path to tree file or loaded in tree (class = phylo)}

\item{og}{Optional: character string of the name of the outgroup (has to
match what it is called in the tree)}

\item{remove_multi_annots}{Logical flag indicating if you want to remove
rows with multiple annotations - alternative is to split rows with mutliple
annotations (default = FALSE)}

\item{return_binary_matrix}{Logical flag indicating if you want to return a
binary matrix (default = TRUE)}

\item{ref_to_anc}{Whether to reference to the ancestral allele to create the binary marix (default = TRUE)}

\item{keep_conf_only}{Logical flag indicating if only confident variants should be kept (1's in Ali's pipeline, otherwise 3's are also kept) (default = TRUE)}

\item{mat_suffix}{Suffix to remove from code and allele matrices so the names match with the tree tip labels.}
}
\value{
list of allele mat, code mat, binary mat and corresponding parsed
  annotations. output will depend on arguments to the function.
}
\description{
Input matrices generated from internal (Ali's) variant calling
  pipeline. Always returns parsed annotation info. In addition, you have the
  option to: 1. split rows with multiple annotations (snps in overlapping
  genes, multiallelic snps) 2. Re-reference to the ancestral allele at that
  position (instead of to the reference genome) 3. Simplify the code matrix -
  which contains numbers from -4 to 3 indicating different information about
  the variants - to a binary matrix indicating simple presence/absence of a
  SNP at that site.
}
