% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdif_data.R
\docType{data}
\name{parsed_indel_cdif3}
\alias{parsed_indel_cdif3}
\title{Parsed indel matrices 2}
\format{
list
Parsed indel matrices 3

list
}
\usage{
parsed_indel_cdif3
}
\description{
Parsed indel matrices 2
}
\keyword{datasets}
