% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdif_data.R
\docType{data}
\name{parsed_indel_cdif8}
\alias{parsed_indel_cdif8}
\title{Parsed indel matrices 6}
\format{
list
Parsed indel matrices 7

list
Parsed indel matrices 8

list
\describe{
  \item{parsed_indel_cdif8}{
  parse_indels(varmat_code = cdif_indel_code_mat,
                             varmat_allele = cdif_indel_allele_mat,
                             tree = cdif_tree,
                             og = NULL,
                             remove_multi_annots = FALSE,
                             return_binary_matrix = FALSE,
                             ref_to_anc = FALSE,
                             keep_conf_only = FALSE,
                             mat_suffix = "_R1_001.fastq.gz")}
 }
}
\usage{
parsed_indel_cdif8
}
\description{
Parsed indel matrices 6
}
\keyword{datasets}
