% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{remove_NA_rows}
\alias{remove_NA_rows}
\title{Remove any rows in the matrix that are NAs}
\usage{
remove_NA_rows(mat, annot_mat, reref_vec)
}
\arguments{
\item{mat}{Referenced binary matrix}
}
\value{
The same or a subset of the binary matrix.
}
\description{
This situation arises in the rereferencing step when it's a
  "complicated" multiallelic site. In the future we could do something more
  nuanced so as to not assign NAs in the first place, but for now just giving
  it NA and removing them.
}
