% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{remove_rows_with_multiple_annots}
\alias{remove_rows_with_multiple_annots}
\title{Remove any rows with multiple annotations}
\usage{
remove_rows_with_multiple_annots(varmat)
}
\arguments{
\item{varmat}{- data.frame where the rows are variants, the columns are genomes,
and the row.names are annotations}
}
\value{
Returns a varmat (class = data.frame) with all rows with multiple annotations
removed. Also writes a file called YEAR_MONTH_DATE_rows_with_multiple_annots_removed
indicating which rows were removed.
}
\description{
- Bypass dealing with rows with multiple annotations (due to
overlapping genes or multiallelic sites) by removing them from the data.frame.
Useful especially as we are testing this function and the functionality to deal
with sites with multiple annotations is not ready.
}
