% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{remove_rows_with_no_variants_or_completely_masked}
\alias{remove_rows_with_no_variants_or_completely_masked}
\title{Remove rows with no variants or that are completely masked}
\usage{
remove_rows_with_no_variants_or_completely_masked(varmat_code,
  varmat_allele)
}
\arguments{
\item{varmat_code}{- data.frame where the rows are variants (numeric description
variants: numbers ranging from -4 to 3), the columns are genomes, and the
row.names are annotations}

\item{varmat_allele}{- data.frame where the rows are variants (character description
variants: A,C,T,G,N,-), the columns are genomes, and the row.names are annotations}
}
\value{
Returns a list with the following elements (in order):
1. varmat_code (class = data.frame) with non-variant rows removed. All rows have at least
one sample with a variant.
2. varmat_allele (class = data.frame) with non-variant rows removed. All rows have at least
one sample with a variant.
1 and 2 should be the same dimensions and have the same row names.
Also writes a file called YEAR_MONTH_DATE_rows_removed_because_no_variants
logging the row names of the removed rows.
}
\description{
Removes rows that have no variants (the same allele in every
sample +/- N and dash) or rows that are completely masked (all Ns).
}
