% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{remove_snps}
\alias{remove_snps}
\title{Remove any row from an indel matrix that has the word SNP in the annotation}
\usage{
remove_snps(mat)
}
\arguments{
\item{mat}{Indel matrix. Rows = indels, columns = samples.}
}
\value{
Indel matrix
}
\description{
I found a row in one indel matrix that called a SNP, ex:
"No_protein_coding/intergenic_region_field_in_ANN SNP at 31494 > C
functional=NULL_NULL_NULL locus_tag=JBHKLKBP_00048 Strand Information:
JBHKLKBP_00048=+;C" - so this function will simply remove this from the indel
parsing pipeline.
}
