% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{remove_unknown_anc}
\alias{remove_unknown_anc}
\title{Remove unknown ancestral states}
\usage{
remove_unknown_anc(varmat_code, varmat_allele, annots)
}
\arguments{
\item{annots}{}
}
\description{
Remove rows from variant matrix where the ancestral state is unknown (- or N)
}
