% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{report_mat_type}
\alias{report_mat_type}
\title{Standardize matrix type input character to all upper case}
\usage{
report_mat_type(matrix_type)
}
\arguments{
\item{matrix_type}{Character ("SNP" or "INDEL")}
}
\value{
matrix type Character ("SNP" or "INDEL")
}
\description{
Standardize matrix type input character to all upper case
}
