% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{root_tree_og}
\alias{root_tree_og}
\title{Root tree on outgroup}
\usage{
root_tree_og(tree, outgroup = NULL)
}
\arguments{
\item{tree}{phylogenetic tree or file path of tree}

\item{outgroup}{tip name of outgroup in phylogeny. If NULL, midpoint root if
not rooted}
}
\value{
rooted tree without outgroup
}
\description{
Root tree based on outgroup. If outgroup is null and the tree
  isn't rooted, midpoint root tree. If tree is rooted, return tree as is.
}
