% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{standardize_row_and_col_names}
\alias{standardize_row_and_col_names}
\title{Add names and standardize row names for input matrices}
\usage{
standardize_row_and_col_names(mat, suffix)
}
\arguments{
\item{mat}{Matrix}
}
\value{
Matrix with updated names and row.names()
}
\description{
Add names and standardize row names for input matrices
}
