% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_clustering_on_tree.R
\name{get_largest_subtree}
\alias{get_largest_subtree}
\title{Get largest pure subtrees}
\usage{
get_largest_subtree(subtrs, isolate_labels, control_labels = NULL)
}
\arguments{
\item{subtrs}{Subtrees created using ape::subtrees to look for clustering on. Should include all isolates of interest.}

\item{isolate_labels}{Named vector of labels by which pure clusters are defined. Names must be equivalent to tree tip label names.}

\item{control_labels}{Named vector of labels known to cluster. Names must be equivalent to tree tip label names. This controls for clustering by requiring that the pure clusters must contain multiple of the control labels.}
}
\value{
list containing the largest pure subtree that each isolate belongs to and the index of that subtree
}
\description{
Get largest pure subtrees
}
